API/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.28.35
botocore==1.31.35
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-guardian==2.4.0
django-rest-knox==4.2.0
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
msgpack==1.0.5
packaging==23.1
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==5.0.0
s3transfer==0.6.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
zope.interface==6.0
